refactor notify into _sendRequest
