+++ lib/airbrake.js
@@ -99,13 +99,8 @@ Airbrake.prototype.log = function(str) {
   if(this.consoleLogError) console.error(str);
 };
 
-Airbrake.prototype.notify = function(err, cb) {
+Airbrake.prototype._sendRequest = function(err, cb){
   var callback = this._callback(cb);
-  // log errors instead of posting to airbrake if a dev enviroment
-  if (this.developmentEnvironments.indexOf(this.env) != -1) {
-    this.log(err);
-    return callback(null, null, true);
-  }
 
   var body = this.notifyXml(err);
 
@@ -151,6 +146,17 @@ Airbrake.prototype.notify = function(err, cb) {
 
     callback(null, url);
   });
+}
+
+Airbrake.prototype.notify = function(err, cb) {
+  var callback = this._callback(cb);
+  // log errors instead of posting to airbrake if a dev enviroment
+  if (this.developmentEnvironments.indexOf(this.env) != -1) {
+    this.log(err);
+    return callback(null, null, true);
+  }
+
+  return this._sendRequest(err, callback);
 };
 
 Airbrake.prototype._callback = function(cb) {
